+++ iscc_schema/__init__.py
@@ -1,2 +1,2 @@
 __version__ = "0.5.0"
-from iscc_schema.schema import IsccMeta
+from iscc_schema.schema import IsccMeta, Signature